pkg/endpoint: add configuration option-related code to options.go
Signed-off by: Ian Vernon <ian@cilium.io>
